Build route
